UnboundedMailbox: simple fix to avoid goroutine overhea
To avoid destroying and recreating goroutines for handling messages,
in cases where the inter-arrival time of messages is not short
enough or long enough to prevent the goroutine from exiting, it is
better to simply wait and check the queues a few time before letting
go of the goroutine. This adds a bit of CPU utilization and possibly
small memory overhead, but boost the performance for those corner
cases almost 3 times.