ConsistentHashRouter: race condition
The consistent hash router suffered from concurrent map read and
write problem. Moreover, since hash ring and hashmap are being
updated separately, there is a race condition where node is selected
from the old hashring but passed to the new hashmap. When a node is
removed, this can cause messages to be routed to a nonexistent node.

A container is introduced to wrap these variables. Updating it only
happens after the new structures are created and the pointer is
updated at once. Also, a local copy of the pointer is made upon
reading from the container to avoid the second problem.