routeractor: Removed node needs to be stopped When a node is removed from the routees list, it should be stopped, otherwise it will remain in the system without having any messages to process.
However, the stop messages should be sent with a delay to give the
removed node to be able to process the messages in its mailbox
(similar to how AKKA does it). Currently, a goroutine is spawned to
stop the actor after a timeout, but this needs to be fixed when
alternative ways are available.