Fixed ConsistentHashRouter and modified router actor (#75)
* ConsistentHashRouter: race condition

The consistent hash router suffered from concurrent map read and
write problem. Moreover, since hash ring and hashmap are being
updated separately, there is a race condition where node is selected
from the old hashring but passed to the new hashmap. When a node is
removed, this can cause messages to be routed to a nonexistent node.

A container is introduced to wrap these variables. Updating it only
happens after the new structures are created and the pointer is
updated at once. Also, a local copy of the pointer is made upon
reading from the container to avoid the second problem.

* ConsistentHashRouter: test concurrency

This test adds and removes multiple nodes, and route messages all
at the same time to test how well the system behave under concurrent
accesses.

* routeractor: Removed node needs to be stopped
When a node is removed from the routees list, it should be stopped,
otherwise it will remain in the system without having any messages
to process.

However, the stop messages should be sent with a delay to give the
removed node to be able to process the messages in its mailbox
(similar to how AKKA does it). Currently, a goroutine is spawned to
stop the actor after a timeout, but this needs to be fixed when
alternative ways are available.

* routeractor: replaced Stop with sending a poison pill