Refactor the middleware construction
Inbound and outbound middleware chains are now built not in props, but at the moment of a local context creation. It allows capturing the local context in a closure and using it in a chain. The immediate benefit is that a context interface value can now be passed through the inbound middleware chain till the very end without asserting it to any particular type. Any middleware can now substitute the context value and pass it to the next element of the chain.