FromTemplate: produce an actor that resets correctly
The new function FromTemplate can be used to create actors by copying from
an instance (known as the template for the actor). Each time the actor is
restarted, its state is reset to the contents of the template. This is
more intuitive than FromInstance, which should be reserved for immutable
value-type actors (and has been documented as such).