Fix consul member re-registration after shutdown #377
When the cluster was shut down and the member deregistration resolved while the health check TTL was
updated, the member was re-registered and as a result was displayed in
consul for 1 minute with the status 'crititcal'.
A re-registration could also happen when the member's status value was
updated after the shutdown finished.